added Vehicle models
