--- C#/Car.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace TollFeeCalculator
-{
-    public class Car : Vehicle
-    {
-        public String GetVehicleType()
-        {
-            return "Car";
-        }
-    }
-}
