--- C#/Motorbike.cs
@@ -1,10 +1,0 @@
-﻿namespace TollFeeCalculator
-{
-    public class Motorbike : Vehicle
-    {
-        public string GetVehicleType()
-        {
-            return "Motorbike";
-        }
-    }
-}
